Mesh positions f32 -> f64; f32 survives only at GPU upload
Kernel round-trips Manifold via MeshGL64 (store boundary no longer
quantizes/re-welds), op_solid_from_mesh takes Float64Array, vendored
three generators emit Float64BufferAttribute positions, canonical mesh
hashing writes f64 bits (FORMAT_VERSION 5). Render/engine reads drop
their f32 up-casts; the one remaining f32 conversion is per-mesh GPU
vertex upload. f32-regression tripwire tests (bit-exact 0.1 / near-1e7
probes) pin every seam: odm-ir/odm-kernel/odm-render precision tests +
conformance three-f64.js; each verified to fail on a simulated
regression. Golden IR hashes regenerated. Plan mesh-f64.md executed and
folded into notes.

+++ crates/odm-kernel/src/diagnose.rs
@@ -1,7 +1,7 @@
 /// Agent-readable diagnosis for meshes Manifold rejects. Manifold's own error
 /// is a bare `NotManifold`; we inspect the (already merge()d) index buffer to
 /// say why.
-pub fn diagnose_open_mesh(vert_properties: &[f32], tri_verts: &[u32], raw_err: &str) -> String {
+pub fn diagnose_open_mesh(vert_properties: &[f64], tri_verts: &[u64], raw_err: &str) -> String {
     use std::collections::HashMap;
 
     let verts = vert_properties.len() / 3;
@@ -9,7 +9,7 @@ pub fn diagnose_open_mesh(vert_properties: &[f32], tri_verts: &[u32], raw_err: &
 
     // Count undirected edge usage: a closed manifold uses every edge exactly
     // twice (once per direction).
-    let mut edges: HashMap<(u32, u32), u32> = HashMap::new();
+    let mut edges: HashMap<(u64, u64), u32> = HashMap::new();
     for t in tri_verts.chunks_exact(3) {
         for (a, b) in [(t[0], t[1]), (t[1], t[2]), (t[2], t[0])] {
             let key = (a.min(b), a.max(b));
